py/pyproject.toml
[project]
name = "led"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pysdl2>=0.9.17",
]
